Create passwordgenerator
Having trouble trying to encrypt the password after it shows you a new random password.  Where I call the password.encode I get a AttributeError: 'list' object has no attribute 'encode'. I think it has something to do with it not being in bytes? If so how can I take the list and turn it into a byte after.